Updated Task1 to be in register with Task4
Task4 starts with a robot, as opposed to relying on a Prefab being instantiated.  This is probably not the preferred solution, but it was the short solution for getting LDR and EnvLight interactions to work.  Those programmatic relationships probably just need to be restructured.  In the meantime, rejiggering is taking place so that Robot object can start in a scene, or not.  Task1 is up to date, but 2 and 3 are not.  Furthermore, all of these tasks need trigger events to do things like transition to the next task, score points, or prompt the user for feedback as in the case of Task4 when the user should be prompted to determine the meaning of special wall colors (blue, cyan and green).